Create a UBI-based partner container definition and deployment strategy
For the initial round of CNF testing for Kiknos, Cisco provided a test
partner tool to utilize for CNF-specific testing called IKEster.  In
order to deliver timely results, Ryan overloaded the use of this Pod to
perform generic CNF tests, as IKEster already contained the necessary
dependencies to perform testing (iproute and iputils).

However, now that we are moving on from Kiknos, we need a more generic
test partner mechanism.  This is the initial commit of a more generic
test mechanism.  First, a Dockerfile containing the test-partner
container definition is abstracted.  The test-partner container is
based on Universal Base Image (UBI) version 7.  UBI version 8 does not
work correctly in the New Mexico Laboratory, and is incapable of
installing the iputils dependency (an error occurs).  Thus ubi-7 is
preferred in the initial release.

Second, an extremely basic partner Pod definition is abstracted in
partner.yaml.  The pod is located in the default namespace.

As a note, the "test-network-function" "oc" wrapper will require some
retrofitting with ubi-7, as the shell prompt differs from Ubuntu's.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>