Abstract local test infrastructure
Often, when developing for this project, it became useful to have a live setup
to debug against.  This involved committing, pushing a branch to the remote
origin, then checking out on the other side (i.e., the cluster).  It quickly
became apparent that moving our code to New Mexico by way of GitHub was not a
very well thought out solution for tests that do not require real hardware.

This change introduces local test infrastructure to run Pods locally in a dev
environment (such as minikube).  I have not yet tested with Kind, but there
should be no reason that this code wouldn't work with that either.  Running
"make" will produce two local pods (test and partner).  In order to run the
test-network-function CNF Certification suite against this setup, utilize the
configuration in "local-test-infra/conf.yaml".  This may be a useful way to
do "integration" tests in the future, but is certainly too heavyweight to be
applicable to unit tests.

Additionally, a few other needed dependencies are installed.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>