Add CI Tests for test infra and partner image
Implements three new GitHub Workflows for testing, and a GitHub Action
for setting up a minikube cluster suitable for CI/CD pipelines.

The new `start-minikube` GitHub Action should be considered as the
recommended Action for creating test clusters for TNF resources in CI
workflows, removing the dependency on third-party, untrusted Actions
from the GitHub Marketplace.

The new workflows are:

- `cnf-test-partner-image.yaml`: Tests whether the partner image can be
  built, and whether the image contains all required tools.

- `local-test-infra.yaml`: Tests the OpenShift resources defined in the
  local-test-infra directory.
  The resources are created on the minikube test cluster using commands
  exposed in the Makefile.

- `partner-pod.yaml`: Tests the installation process of the partner pod
  according to the instructions in the README.
  The pod is created on the minikube test cluster.

Signed-off-by: Marek Kochanowski <mkochanowski@redhat.com>