Extend the CI workflow with image push to Quay.io
Introduces a push job to the `cnf-test-partner-image.yaml` workflow.
The new job is dependent on the successful completion of an already
existing job, responsible for building and testing the image.

The image pushed to Quay.io is always built from the `main` branch.

This change extends a CI workflow introduced in CTONET-831.

Signed-off-by: Marek Kochanowski <mkochanowski@redhat.com>