Replace support-tools image with custom one (#67)
* Replace support-tools image with custom one

Using the support-tools image for debug partner pods restricts us to
OpenShift environments in some tests, like diagnostic. We can swap that
one with a UBI8-based image, so those tests can be executed on minikube
or other environments as well.

* Additional debug partner image fixes

- Remove unnecessary REDHAT_RHEL_REGISTRY env var
- In Dockerfile, run yum update before yum install

* Add procps-ng package to debug partner image

It contains the sysctl binary, required for some tests.