Remove test pods runAsUser ID for OCP (#507)
Hardcoding this number for OCP was a bad idea as the valid range might be
different depending on the OCP version. This can be check in the namespace
annotations. This is the valid range for OCP 4.15:
  openshift.io/sa.scc.uid-range: 1000700000/10000

When not provided in the pod's SCC, the user ID is automatically
selected from that range.